requirements.txt
langchain==0.1.12
langchain-openai==0.0.8
langchainhub==0.1.15
beautifulsoup4==4.12.3
tiktoken==0.6.0
faiss-cpu==1.8.0
fastapi==0.110.0
uvicorn==0.28.0
requests==2.31.0